pyproject.toml
[project]
name = "security-data"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "certifi==2025.10.5",
    "cffi==2.0.0",
    "charset-normalizer==3.4.4",
    "coloredlogs>=15.0.1",
    "config>=0.5.1",
    "fastapi>=0.135.1",
    "forex-python>=1.9.2",
    "google==3.0.0",
    "idna==3.11",
    "markdown-it-py==4.0.0",
    "mdurl==0.1.2",
    "multitasking==0.0.12",
    "numpy==2.3.4",
    "pandas==2.3.3",
    "peewee==3.18.2",
    "platformdirs==4.5.0",
    "protobuf==6.33.0",
    "pycparser==2.23",
    "pydantic-settings>=2.13.1",
    "pygments==2.19.2",
    "pyotp>=2.9.0",
    "python-dateutil==2.9.0.post0",
    "python-dotenv>=1.2.2",
    "pytz==2025.2",
    "pyyaml>=6.0.3",
    "requests==2.32.5",
    "rich==14.2.0",
    "six==1.17.0",
    "soupsieve==2.8",
    "typing-extensions==4.15.0",
    "tzdata==2025.2",
    "urllib3==2.5.0",
    "uvicorn[standard]>=0.42.0",
    "websockets==15.0.1",
    "ws-api>=0.33.0",
    "yfinance==1.2.0",
]
